requirements.txt
# -----------------------------
# ChurnGuard Requirements (Python 3.11 Compatible)
# -----------------------------

# Core Framework
streamlit==1.42.0
python-dotenv==1.0.1

# Database
pymongo==4.6.1
motor==3.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.4

# Machine Learning
scikit-learn==1.5.2
joblib==1.4.2

# NLP & Sentiment Analysis
textblob==0.17.1
nltk==3.9.1

# LangChain & Gemini AI
# Removing specific version pins to allow pip to resolve dependency conflicts.
langchain
langchain-core
langchain-community
langchain-experimental
langchain-google-genai
langsmith
google-ai-generativelanguage>=0.7.0,<1.0.0
google-generativeai>=0.8.0

# Visualization
plotly==5.24.1
altair==5.4.1

# Utilities
requests>=2.32.5,<3.0.0
pydantic==2.9.2

# API Framework
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12

# Async Processing
celery==5.4.0
redis==5.2.0
aiohttp==3.10.11

# Email & SMS Services
twilio==9.2.3
boto3==1.35.85

# Monitoring & Logging
structlog==24.4.0
sentry-sdk==2.18.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0